refactor(api)!: idealize the public fprint-* surface
Beauty as the sole criterion: make the published surface one that will not need
an apology after 1.0. No behaviour changes to the kernels — the BOZORTH3/MINDTCT
golden output stays bit-exact.

fprint-core
- DeviceId/DriverId hide their String; construct with `new`, read with `as_str`,
  Display + From<&str>/From<String>. The representation is no longer frozen.
- Growable output structs (DeviceInfo, EnrollProgress, VerifyOutcome,
  IdentifyOutcome, EnrollDate) are `#[non_exhaustive]` with constructors, so a
  future field is not a breaking change. Print gains a builder; fields stay public.
- Error derives Clone/PartialEq/Eq — a comparable, value-semantic vocabulary.
- FingerStatus and Temperature stop being dead speculative surface: Temperature
  is wired onto DeviceInfo, FingerStatus onto EnrollProgress. The libfprint shim
  is their real producer (raw fp_device_get_temperature / _finger_status FFI,
  tracked as M2-E); the native virtual device reports None/NONE honestly.

fprint-mindtct
- GrayImage validates on construction: `GrayImage::new(..) -> Result<_, ImageError>`
  with private fields and accessors, so detect_minutiae can no longer panic on a
  short buffer — the panic became a typed Err.
- The diagnostic surface (DebugMaps, RawMinutia, debug_*) moves behind the
  default-off `unstable-diagnostics` feature, out of the stable semver contract.
- The two module-level too_many_arguments allows narrow to per-function.

fprint-bozorth3
- match_score is now total: the squared-distance/angle computations widen to i64,
  so it neither panics under debug_assertions nor silently wraps in release on
  extreme coordinates. Valid-reader scores are unchanged (golden score-exact).

The rest of the tree migrates to these constructors (backends, fp3, integration,
cli, driverkit, daemon, tests, doctests). fp3 also stops re-exporting its `Result`
alias so a glob-import beside fprint-core does not clash.

+++ crates/fprint-bozorth3/src/cluster.rs
@@ -590,8 +590,10 @@ impl Bz {
             let cj = self.avv[tpu][2] - self.avv[iiu][2];
             let ck = self.avv[tpu][3] - self.avv[iiu][3];
             let cji = self.avv[tpu][4] - self.avv[iiu][4];
-            let tt = (cl * cl + cj * cj) as f32;
-            let ai = (cji * cji + ck * ck) as f32;
+            // Squared in i64 so far-apart centroids cannot overflow the multiply; the sum feeds an
+            // f32, and for in-range coordinates the widened value is the same integer.
+            let tt = (i64::from(cl) * i64::from(cl) + i64::from(cj) * i64::from(cj)) as f32;
+            let ai = (i64::from(cji) * i64::from(cji) + i64::from(ck) * i64::from(ck)) as f32;
             let fi2 = (2.0_f32 * TK) * (tt + ai);
             let dz = tt - ai;
             if dz * dz > fi2 * fi2 {
